refactor: 체결 로직 리팩토링 및 테스트 코드 일부 작성
