+++ src/main/java/com/cleanengine/coin/order/application/OrderService.java
@@ -2,7 +2,14 @@ package com.cleanengine.coin.order.application;
 
 import com.cleanengine.coin.order.application.dto.OrderCommand;
 import com.cleanengine.coin.order.application.dto.OrderInfo;
+import com.cleanengine.coin.common.error.BusinessException;
+import com.cleanengine.coin.common.response.ErrorStatus;
+import com.cleanengine.coin.order.adapter.out.persistentce.order.command.BuyOrderRepository;
+import com.cleanengine.coin.order.adapter.out.persistentce.order.command.SellOrderRepository;
 import com.cleanengine.coin.order.application.strategy.CreateOrderStrategy;
+import com.cleanengine.coin.order.domain.BuyOrder;
+import com.cleanengine.coin.order.domain.Order;
+import com.cleanengine.coin.order.domain.SellOrder;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,6 +28,8 @@ import static com.cleanengine.coin.common.CommonValues.SELL_ORDER_BOT_ID;
 @Validated
 public class OrderService {
     private final List<CreateOrderStrategy<?, ?>> createOrderStrategies;
+    private final BuyOrderRepository buyOrderRepository;
+    private final SellOrderRepository sellOrderRepository;
 
     @Transactional
     public OrderInfo<?> createOrder(@Valid OrderCommand.CreateOrder createOrder){
@@ -38,5 +47,28 @@ public class OrderService {
                 false, orderSize, price, LocalDateTime.now(), true);
 
         return createOrder(createOrder);
+    }
+
+    @Transactional
+    public Order updateOrder(Order order){
+        /*
+        CreateOrderStrategy<?, ?> strategy = createOrderStrategies.stream()
+                .filter(s -> s.supports(order instanceof BuyOrder))
+                .findFirst()
+                .orElseThrow(() -> new BusinessException(
+                        "Unsupported order type: " + order.getClass().getName(),
+                        ErrorStatus.INTERNAL_SERVER_ERROR)
+                );
+        strategy.saveOrder(order);
+        return order;
+        */
+
+        if (order instanceof BuyOrder) {
+            return buyOrderRepository.save((BuyOrder) order);
+        } else if (order instanceof SellOrder) {
+            return sellOrderRepository.save((SellOrder) order);
+        } else {
+            throw new BusinessException("Unsupported order type: " + order.getClass().getName(), ErrorStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
