+++ src/main/java/com/cleanengine/coin/trade/application/TradeBatchProcessor.java
@@ -1,12 +1,10 @@
 package com.cleanengine.coin.trade.application;
 
-import com.cleanengine.coin.chart.dto.TradeEventDto;
 import com.cleanengine.coin.order.domain.spi.WaitingOrdersManager;
-import com.cleanengine.coin.orderbook.application.service.UpdateOrderBookUsecase;
 import jakarta.annotation.PreDestroy;
 import lombok.Getter;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
@@ -21,27 +19,20 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
-@Service
 @Order(4)
+@Slf4j
+@RequiredArgsConstructor
+@Service
 public class TradeBatchProcessor implements ApplicationRunner {
 
-    Logger logger = LoggerFactory.getLogger(TradeBatchProcessor.class);
-
     private final WaitingOrdersManager waitingOrdersManager;
-    private final TradeService tradeService;
+    private final TradeFlowService tradeFlowService;
     private final List<ExecutorService> executors = new ArrayList<>();
 
     @Getter
     private final Map<String, TradeQueueManager> tradeQueueManagers = new HashMap<>();
-    private final UpdateOrderBookUsecase updateOrderBookUsecase;
 
     @Value("${order.tickers}") String[] tickers;
-
-    public TradeBatchProcessor(WaitingOrdersManager waitingOrdersManager, TradeService tradeService, UpdateOrderBookUsecase updateOrderBookUsecase) {
-        this.waitingOrdersManager = waitingOrdersManager;
-        this.tradeService = tradeService;
-        this.updateOrderBookUsecase = updateOrderBookUsecase;
-    }
 
     @Override
     public void run(ApplicationArguments args) {
@@ -51,8 +42,7 @@ public class TradeBatchProcessor implements ApplicationRunner {
     private void processTrades() {
         for (String ticker : tickers) {
             TradeQueueManager tradeQueueManager = new TradeQueueManager(waitingOrdersManager.getWaitingOrders(ticker),
-                    updateOrderBookUsecase,
-                    tradeService);
+                    tradeFlowService);
             tradeQueueManagers.put(ticker, tradeQueueManager); // 정상 종료를 위해 저장
 
             ExecutorService tradeExecutor = Executors.newSingleThreadExecutor(r -> {
@@ -66,7 +56,7 @@ public class TradeBatchProcessor implements ApplicationRunner {
                 try {
                     tradeQueueManager.run();
                 } catch (Exception e) {
-                    logger.error("Error in trade loop for {}: {}",ticker, e.getMessage());
+                    log.error("Error in trade loop for {}: {}",ticker, e.getMessage());
                 }
             });
         }
@@ -87,7 +77,7 @@ public class TradeBatchProcessor implements ApplicationRunner {
                     executor.shutdownNow();
                     // 추가로 1초 더 대기
                     if (!executor.awaitTermination(1, TimeUnit.SECONDS)) {
-                        System.err.println("스레드풀이 완전히 종료되지 않았습니다");
+                        log.error("스레드풀이 완전히 종료되지 않았습니다");
                     }
                 }
             } catch (InterruptedException e) {
@@ -95,11 +85,6 @@ public class TradeBatchProcessor implements ApplicationRunner {
                 Thread.currentThread().interrupt();
             }
         }
-    }
-
-    @Deprecated
-    public TradeEventDto retrieveTradeEventDto(String ticker) {
-        return null;
     }
 
 }
