+++ src/main/java/com/cleanengine/coin/trade/application/TradeExecutedEvent.java
@@ -7,7 +7,21 @@ import lombok.Getter;
 @Getter
 @Builder
 public class TradeExecutedEvent {
+
     Trade trade;
+
     Long buyOrderId;
+
     Long sellOrderId;
+
+    private TradeExecutedEvent(Trade trade, Long buyOrderId, Long sellOrderId) {
+        this.trade = trade;
+        this.buyOrderId = buyOrderId;
+        this.sellOrderId = sellOrderId;
+    }
+
+    public static TradeExecutedEvent of(Trade trade, Long buyOrderId, Long sellOrderId) {
+        return new TradeExecutedEvent(trade, buyOrderId, sellOrderId);
+    }
+
 }
