+++ src/main/java/com/cleanengine/coin/trade/application/TradeQueueManager.java
@@ -1,7 +1,6 @@
 package com.cleanengine.coin.trade.application;
 
 import com.cleanengine.coin.order.domain.spi.WaitingOrders;
-import com.cleanengine.coin.orderbook.application.service.UpdateOrderBookUsecase;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -10,10 +9,10 @@ public class TradeQueueManager {
     private volatile boolean running = true; // 무한루프 종료 플래그
 
     private final String ticker;
-    private final TradeService tradeService;
+    private final TradeFlowService tradeFlowService;
 
-    public TradeQueueManager(WaitingOrders waitingOrders, UpdateOrderBookUsecase updateOrderBookUsecase, TradeService tradeService) {
-        this.tradeService = tradeService;
+    public TradeQueueManager(WaitingOrders waitingOrders, TradeFlowService tradeFlowService) {
+        this.tradeFlowService = tradeFlowService;
         this.ticker = waitingOrders.getTicker();
     }
 
@@ -21,7 +20,7 @@ public class TradeQueueManager {
         // TODO : 주문 시 이벤트 기반으로 동작하도록 개선
         while (running) {
             try {
-                tradeService.execMatchAndTrade(ticker);
+                tradeFlowService.execMatchAndTrade(ticker);
             } catch (Exception e) {
                 log.error("Error processing trades for {}: {}", this.ticker, e.getMessage());
             }
