+++ src/main/java/com/cleanengine/coin/user/info/application/AccountService.java
@@ -4,6 +4,8 @@ import com.cleanengine.coin.user.domain.Account;
 import com.cleanengine.coin.user.info.infra.AccountRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class AccountService {
 
@@ -15,6 +17,14 @@ public class AccountService {
 
     public Account retrieveAccountByUserId(Integer userId) {
         return accountRepository.findByUserId(userId).orElse(null);
+    }
+
+    public Optional<Account> findAccountByUserId(Integer userId) {
+        return accountRepository.findByUserId(userId);
+    }
+
+    public Account save(Account account) {
+        return accountRepository.save(account);
     }
 
     public Account createNewAccount(Integer userId, double cash) {
