+++ src/test/java/com/cleanengine/coin/trade/application/TradeServiceTest.java
@@ -1,0 +1,7 @@
+package com.cleanengine.coin.trade.application;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class TradeServiceTest {
+
+}
